Make minor changes to create a better Smart demo.
Clean up Smart/index.act and demo.act files (and variations like
smartDemoIndex.act) to make them clearer for demos and production.

Change several table settings to make search the initial action rather
than displayAll (search will be more appropriate as the tables get
bigger, although displayAll was convenient for quicker testing).

Modify TableController's add action so that it does a search after an
add rather than a record view, since there now might be multiple records
matching the provided record identifying information.

Fix bug -- ModuleEnrollment search did not work on Instructor preferred
first name (needed alias).

Fix bug -- Modify Acl.php so that when a resource is requested that does
not exist, Smart reports that the user is not authorized to access that
resource rather than that the resource does not exist.  ("Not
authorized" is a better end-user message, although "resource does not
exist" is an informative message for a developer or DBA.)