Add models tests; New selectFrom table setting attribute
Create or polish test suites for application/models classes:
    RampIniReader (complete)
    ActivitySpec (complete)
    ActivityGateway (complete)
    TVSGateway (complete)
    DbTableValidValuesSource (complete)
    DbTableTable (complete)
    Field (complete, except need to develop better messages for 2 exceptions)
    SetTable (mostly done)
    Note: Test suites still needed for TableViewSequence, TVSFactory,
        and ExternalTableRef in the models directory.

Add support for a new selectFrom attribute in table settings, to be used
when the set of valid values for a particular table field is defined in
another table (e.g., a table listing valid terms, departments, or majors).

Clean up some documentation comments and exception messages in several
classes in application/models.