Implement authentication with encrypted passwords
Implement authentication with encrypted passwords.  New users will now
be created with default passwords, causing Ramp to send them to a Set
Password screen the first time they try to log in.  Users may log in
explicitly from the menu bar, or will be redirected to the Log In
screen if they attempt an action for which they are not authorized.

Modify Table controller to prohibit adding or modifying passwords to the
Users table using standard RAMP table actions.  Provide a new Change
Password option on the menu bar when users are logged in which will go
to a separate screen just for that purpose, ensuring that new passwords
will be encrypted correctly.  Provide an administrative screen to reset
user's passwords back to the default if they have forgotten their
password (Reset Password).

Move all table settings related to the Ramp and Smart demos to a
separate demoSettings directory and move other, development-specific
settings/actions/menus (including many test settings) to a separate
devSettings directory.  These moves clear the settings directory for
an actual Ramp/Smart application.

Rename the ManageAuths settings directory to Admin, in anticipation of
new administrative tasks not related to the Users or Authorizations
tables (e.g., removing zombie locks).

Improve the installation instructions and procedures in several ways:
  -- Create a section specifically about addressing security concerns.
  -- Create scripts and special instructions allowing for the
     installation of the Ramp demo or Smart demo by running just a
     single MySQL script (after addressing a few security issues).
  -- Create simpler scripts that can be used as guides for installing
     customized Ramp/Smart applications other than a demo.
  -- Begin creating an installation document that will explain how to
     customize the application/configs/application.ini file.
  -- Move the configuration instructions for the Ramp and Smart
     demos to separate files to remove the expectation that new
     Ramp/Smart applications would include demo environments.

Commit several small changes that arose from previous testing of
TableViewSequence, affecting TableViewSequence, TVSGateway, and
SetTable.