Support Markdown in documents and activity descriptions
Add support for displaying text, HTML, and Markdown documents (new
'document' activity type).

Add support for Markdown in documents and in activity list comments
and button descriptions.

Add document resources to list of resources that can be stored in
ramp_auth_auths table and to list of resources that are extracted from
that table.

Add support for a documentationRoot property in application.ini.
Create a new application/docs directory and set the documentationRoot
property in all the various sample configuration files in the configs
directory to be the new directory.  Move all Ramp and Smart
documentation (aboutRamp, aboutSmart, and some drafty documentation
that was in a ramp/docs directory) to a new rampDocs directory under
the documentation root (application/docs/).

Introduce Ramp_Controller_KeyParameters class to eliminate need to
repeat code to get the key parameter for various types of controllers in
the init methods of those controllers and in the ACL plugin (which needs
the various key parameter to determine the resource being requested).

Make a number of functions in Ramp_Ini_reader static, and make
getSettingsDirectory method public.

Remove html and markdown activity types because they're too much like
comments and we could always look for Markdown markup in comments ?