Add support for base menu directory
Add support for a base menu directory so that default menu filename and
role-based menu filenames may be specified as either full pathnames or
relative pathnames (relative to the base menu directory).  Update sample
configuration files accordingly.

Add notes in README.txt and INSTALL.txt that Ramp requires PHP 5.3.
This was not true originally, and was tested at some point with
5.2.something, but Ramp now frequently uses the variation on the
ternary operator, new in 5.3, which allows the middle section to be left
out.

Improve description of Report activity in rampDemo/index.act.