Support "inactive" property for activity specs
Support "inactive" property for activity specifications, which causes
the button for that activity to become inactive (and changes in the
appearance of the button & description, depending on css style
specifications).

Remove menu options from Smart Demo guest menu that guests are not
authorized to do.  In main Smart Demo guest activity list, inactivate
activities that guests are not allowed to do and indicate which Demo
username has authorization for those activities.

Clean up the guest activity list in the docs/rampDocs directory to show
only those documents (or initial notes for documents) that have some
relevance to Demo users.  Create a second docs/rampDocs activity list
for developers that lists more documents (currently document fragments
and placeholders).