Clean up Demo and Development environments
Bring installation/SmartDevSetup and application/devSettings/Smart
directories up to date with installation/SmartDemoSetup and
application/demoSettings/Smart directories.

Add authorizations in the development environment to modify tables
that are read-only in the demo environment.

Update role-based menus for Smart demo and development environments.
Create variations on activity lists based on whether user is logged in
or not (and logged in playing which role) to reduce the number of
activities being presented that the user is not authorized to do.  Make
some activities "inactive" to indicate that user does not have
authorization or that activity has not yet been implemented.

Update sample application.ini template files to match changes to menus
and activities and also to make better use of section inheritance.

Make sub-headings in activity lists bold to distinguish them from
comments.