Create Ramp_RegistryFacade facade to Zend_Registry
Create Ramp_RegistryFacade to provide functions that return various
configuration settings, or default values for configuration settings that
have not been provided.  This provides a buffer between the hard-coded
property names from the application.ini configuration file and the
various classes that use those property settings.  Update classes that
had been using Zend_Registry with hard-coded property names.

Remove earlier version of the RegistryFacade (RampConfigs), which was in
models directory.  Put new facade in library, so that in the future
it can be called within the Bootstrap.php file (for menu setup).