Clean up look & feel of activity, list, and table pages
Clean up the layout, fonts, colors, etc of the Activity List screen.
Establish the look and feel of the consistent Ramp/Smart header and
menu.  Clean up the colors and fonts on the list and table view screens,
including hover behavior.

Still need to clean up the fonts, colors, etc of the single-record views
(view, add, edit, delete, and search).