Make it easier to create related records
Resolves  #19 and #26

Provide three methods to make it easier to create related records (e.g.,
new Student record having just created Person record, new ModuleOffering
record from, or referring to, existing Module):

  - __"Drill-down" functionality:__
Add External References to Staff and Student from Person to provide
a "drill-down" option with search.  If we have created a new Person
entry and now need to make the person a Student, we click on Student
Record external reference which will come up with no matches found
and then we can Add New Entry with the foreign key automagically
filled in.  (This is not new functionality, actually, but new
understanding and illustration of existing code that provides
seemingly new functionality.)

  - __"Add-another" functionality:__
Improve the "Add New Entry" button on list-view and table-view so that
when the list/table is a subset based on a search or filter, the new
entry will have the search/filter values automagically filled in.  (Can
be changed if the new entry is not, in fact, related to the previous
search or filter.)  Example: clicking on the Module Offerings external
reference of a Module, or searching for that module's ID or code/number
explicitly, results in a list of existing Offerings for that Module (or
an empty list).  Choosing Add New Entry will create a new Module
Offering record with the Module ID already filled in.

  - __"Look up key" functionality:__
Move link to "selectUsing" tables to a magnifying glass next to text
input field rather than as link on field label.  This allows foreign key
lookup to work well, which it did not when it was a link on the label.

Fix getting imported table names for authorization.

Fix bug with decorating Hidden Elements caused by refactoring of
TableRecordEntry form.

Add support for separate deleteSetting property in table settings.

Change name of "advisor" field in Student table to "advisorID"
(demoSettings and devSettings).

Clean up Person.ini, Student.ini, using inherited sections (demoSettings
and devSettings).

Remove remaining @version javadoc tags (which were completely inaccurate
anyway), relying on git versioning instead.