Add support for split-view table views
Add support for split-view table views when a search results in a number
of records with some shared data fields.  For example, if a search in an
enrollment history table focuses on a single student, the basic
identifying information for the student (which is the same for all
records) appears once in the top of the screen, while the student's
enrollment history appears in a list in the bottom half of the screen.

TODO: Could use a lot of style work to look much better.

This feature is a step towards providing block entry of data.