Support block data entry of new records.
Closes #23

Add new feature that allows user to go from a split view of multiple
table records (possible when the records being viewed have some field
values in common) to a block data entry screen that allows the user to
add multiple new records having the same field values in common.  For
example, a split view of a class list would show the basic class
information once, at the top of the screen, followed by a list of all
the students enrolled in the class.  (Requirement: there must be at
least two students enrolled in the class to establish which fields are
in common and which are not for the split view.)  The associated block
record entry screen would display the same class information at the top
and the same summary of enrolled students at the bottom, but would
provide a set of empty text fields to the right of the class information
in which the user can enter new student IDs.