Add support for LDAP authentication (untested)
Add initial support for LDAP-based authentication (e.g., Active
Directory or OpenLDAP), although the new code is not yet tested nor
active (code throws an exception if the authentication type in
application.ini is anything other than "internal").

For internal authentication, remove the specification of default
password in application.ini, relying on the default provided in the
schema for ramp_auth_users instead.